Have `playpass create` handle existing projects, and update docs.
If you run `playpass create` in an existing project with a package.json,
you'll be prompted to add Playpass to it instead of creating from a
template.

Choosing "yes" will reserve a game ID, generate a `playpass.toml`, and
print out some instructions of how to modify your game JS to init
playpass.