Implement sharing on desktop.
Preview: https://share--kitchen-sink.playpass.games

Blacklist native web sharing on desktop because the UX is not great in
the one browser it's currently supported in (Windows Chrome).

This popup will also be displayed on mobile if navigator.share isn't
available for some reason, such as running on a non-HTTPS site.